Parametrize patch strip option in workarounds.
Right now, all the workarounds are patched using the strip option
--p set to 1. That works in most of the cases, but for some directories
/projects it's needed to pass a different strip value (tripleo-ansible)
causing all workarounds to not been applied.

This patch adds a new workarounds option patch_strip which accepts an
integer and will set the -p option at the time of applying the workaround.
If the option is not passed, the default value will be 1.

Change-Id: I0cb5d26838cf38bb04e22cfa3a06e06177eebd55