[train-only] Duplicate rolemap inventory filters in tripleo-upgrade
Related to I3daf77f099fd0277a9d8495932ca3e28ae7b964d, we need to duplicate
the filters in tripleo-upgrade to break a cyclic dependency.

Also remove the tripleo_upgrade prefix from the filter names so that the tasks
do not have to differ on train vs master.
If both tripleo-upgrade and tripleo-ansible are recent train versions it does
not matter which implemenation ansible gives us as they are identical.
On master we only have the tripleo-ansible implementation.

Change-Id: I0fc4f7ef907093af4ba26393a1d59a7fe215a5ee