Simplify ffwd2 CIP stein parameters.
Currently, the ContainerImagePrepare parameters required for FFWD2
to work are harcoded in the defaults/main.yaml. However, the
namespace_stein variable, should point at the same value the namespace
variable points. This patch simplifies a little the logic, adds a new
fact full_namespace that stores the content of the namespace field from
the containers-prepare-parameters.yaml file and assigns the namespace_stein
key to it.

Change-Id: I07967e0590aee9c86a6427469914f65fbdb3929d