Build enablerepo options automatically for leapp when not subsribed.
A new option added into leapp lets us configure a set of repositories to run
leapp with. This new option is compatible with rhos-release, removing the need
for an special leapp actor.
This patch checks the rhos-release repos configured locally in the Undercloud
and builds a --enablerepo string with all the repos obtained. This string is
then passed at the time of running leapp, if the machine isn't using a
subscription.

This patch also cleans up some unnecessary code to handle differente leapp
versions, as our CI now works with latest leapp version. As well as some
unsupported Leapp environment variables [0].

[0]-https://github.com/oamg/leapp/blob/69f3f8c21dae5fbe270701e5728b8e2512faddb6/docs/source/el7toel8/envars.md

Change-Id: Ife2734547c4c981dece0c0f9b436eb10e99dd1b1