Support new Rhel Enforcement in CI.
Starting from 16.0, OSP, the Red Hat version of tripleo, enforce
certain rules between Rhel version and OSP version.

Those checks rely on the Red Hat subscription mechanism.

To support CI when no subscription is apply we need to disable the
enforcement.  This is done by setting SkipRhelEnforcement to True.

So we add code here to disable the check by default for CI.

The new option --enforce-rhel set to yes make it possible to run the
update while keeping the enforcement to test subscribed environment.

Change-Id: Idac7c370aaaed2805787c0b3327bb9b7e91f1561