Remove loop for overcloud_upgrade_roles.yaml.
The overcloud_upgrade_roles.yaml playbook was refactored to take
care of upgrading all the roles within the file. Therefore, there
is no need to iteratively call to it passing oc_roles. If we do,
we would be upgrading the whole environment multiple times.

Also, as the loop is removed we can change the import from an
include_tasks into import_tasks.

Change-Id: Ie2f7263e8c8844fd040dd42755dc010cc70816b1