Add --yes into update commands to autoconfirm.
In [0] a confirmation flag has been included for the updates
procedure to ensure customers are aware of the type of operation
they are about to perform.

We need the --yes flag in tripleo-upgrade so the execution doesn't
get waiting for the user to type yes. This patch will check in the
command's help if the --yes option is available, if so it will
include it in the command execution.

Co-Authored-By: Sergii Golovatouk <sgolovat@redhat.com>

[0] - I838e6748879c668dd004ca2243b7b00b857c2a7b

Change-Id: Ibb0676e7a3d97645e4e8a2553e7ec175e96a14b4
(cherry picked from commit e29de36bfb0cf758990e1762435e0e3d19ab616d)