Remove ffu converge and perform stack update only.
In the Queens to Train upgrade, the CLI command used for fast forward
upgrade is the same one as for upgrades. Changing the converge step to
avoid calling fast-forward-upgrade converge instead. Also, for this
FFU workflow it shouldn't be needed to re-deploy again during the
converge step, so adding the --stack-only flag when
ffu_overcloud_upgrade is set.

Change-Id: Ief897b15749ff606d0ca3a1a1b12882236f30cb9