Add patch revision parameter to workarounds
The current workaround patch mechanism only allows for the
'current' (most recent) version of a patch to be used in
workarounds.

There are times when we may want to pin to a previous patch
set. This change enables that option.

According to [a] the revision ID can be:

- a legacy numeric patch number ("1" for first patch set of the change)
- the literal 'current' to name the current patch set/revision
- a commit ID ("674ac754f91e64a0efb8087e59a176484bd534d1")
- an abbreviated commit ID that uniquely identifies one revision of the change ("674ac754"), at least 4 digits are required

We also clean up two unused workaround templates.

[a] https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#revision-id

Change-Id: I6382d139e123b745d0c23f3afd695dae316d861d
(cherry picked from commit 6fba346ad4c67ab67026e6ad87dca5311af7e92c)