Fix RUN_ANSWER unbount variable issue
Since merging Ibb0676e7a3d97645e4e8a2553e7ec175e96a14b4
"Add --yes into update commands to autoconfirm." into stable/train
my FFU attempts are failing with:
/home/stack/overcloud_system_upgrade-controller-0.sh: line 11: RUN_ANSWER: unbound variable

Reason for this is that RUN_ANSWER gets set only inside the if ceph_osd_enabled
branch, which in my case is false. Let's move setting that variable
outside the if so it gets set everytime.

This script seems to be the only one where it was inadvertently left
inside the if.

Change-Id: I35df29521c1807113710d7d6cc039551400ebe2f