Mesure l3 agent connectivity in seconds instead of percentage.
This make the ping check independent of the task time.  We can now
specify in absolute time how long we tolerate a ping loss during the
ping test.

We set this to 0 by default for update as no such cut should happen
during this workflow.

For other tasks it's 60 seconds.

We also maintain a percentage check for tasks that are very short in
total time (less than 10min for instance) as the total number of
seconds there could be less than 60 seconds, but with a 100% packet
loss.

Eventually, we shield printf against locale that would use a comma as
a separator for number decimal and make sure it's always a point.

Change-Id: I69ac4ee4f7eb3214c925686c1efd1bd2f5972185