Keep ceph-ansible 3 during whole FFU.
To be able to FFU a ceph-enabled environment, it is required to
keep installed in the system ceph-ansible 3 in the Undercloud until the
converge step.

This patch adds the ceph-ansible package in the to_keep files, as well as
modifies some of the Undercloud upgrade tasks to make sure the package
survives to the leapp upgrade.

Co-Authored-By: Jesse Pretorius <jesse@odyssey4.me>
Change-Id: Ide724e398a5cd4fdaf77d1a127e9d5c32852ae6b