Remove skip ceph-validations and --playbook for system_upgrade.
Now that the ceph-systemd validations are fixed, we don't need
to add all the skip ceph validations when running the ceph_systemd
step. Also, for [0] to take action, we need to run all playbooks
so the noout/nobalance params are executed in the
post_upgrades_steps_playbook.yaml. Therefore, the --playbook option
is removed from the system_upgrade step.

[0] - Ib10b6ad50e1052fe69640b39a173e856a426d607

Change-Id: I7c5b81545d56fabd021a306a549485e46a785241