Iterate over all templates which need NW template migration.
There is a bug with the logic that takes care of migrating the
NW templates from one version to another. As it is right now,
it was only performing the NW template migration of the last
file it found a Net::SoftwareConfig in. Causing issues if
more than one template from overcloud_deploy script contained.

This patch loops over all the outputs from nic_config_envs and
for each match it will run the script to migrate the templates.

Change-Id: I1456f55d18d9e87782e0f6d5232bdee7f8a85fb8