Create workload_launch script when sanity testing is enabled
If workload testing is not enabled, then the workload script
is not there, so the sanity test fails.

In this patch we change the conditionals to ensure that the
workload launch script is created if either test is enabled.
we also then ensure that the sanity test launches are using
the correct conditional.

Change-Id: I3d2c7691d43674a0304bea8f366162e808dde3ad
(cherry picked from commit 45c2b65361846b69eac136462f368ecd1b0b76da)