Try to fix our way of spawning scripts
We should always default to executable /usr/bin/bash instead of
adding bash into shell command as it might cause hangs for example
in the timestamper awk waiting for bash and not the script
we wanted to run.

We already set chdir so we don't need to specify {{ working_dir }}.

We should always have in mind reruns and never overwrite already
existing logs. Always use tee -a or >>.

There is discrepancy between need of adding ./ and not for the
scripts. The resolution is that the ones that have path in variable
are having full path in the variable, there is no need to add "./".

Resolves: UPG-2052
Closes-Bug: #1895960
Change-Id: I987c124eecd9d3d906cac7f9ab18ac4fc28165e3
(cherry picked from commit 071cc1d8f0334878c58680868f0bee265ec2c8ac)