Remove _stein _ceph3 parameters before converge.
Leaving the specific FFU parameters before running converge
could be problematic when scaling out or performing any
future operation in the environment.

Adding the code to get rid of them on the FFU automation. As
otherwise the CI job would fail due to the addition of this
extra verification [0] in python-tripleoclient.

[0] - I24715f5e55d4cd6cf9879345980d3a3c5ab8830c

Change-Id: I1d2c206ab46a04b11224b47a3474713ff3525a8f