[Queens Only] Add rhos-release build parameter for UC in FFU
When upgrading the Undercloud in the FFU procedure a sequential
upgrade is carried out. If using rhos-release, the 'latest'
build is used by default, which is not what we want.

This patch extends the ffu_undercloud_repo_args to include the
appropriate default parameters needed for rhos-release and sets
them for the role execution.

This is queens-only due to this whole mechanism being different in
later branches.

Change-Id: Ib763e2350724022d71750f53b39d2652c9853a3c