[queens-only] Allow rhos-release parameters for UC RHEL in FFU
When upgrading the Undercloud in the FFU procedure a sequential
upgrade is carried out. If using rhos-release, the currently
running version of RHEL is used as the default.

This may be fine, but if we want to change that there is currently
no way of doing so.

This patch allows the ffu_undercloud_repo_args to optionally include
the appropriate default parameters needed for rhos-release and sets
them for the role execution. These parameters would be passed to
tripleo-upgrade via an extra var.

This is queens-only due to this whole mechanism being different in
later branches.

We also now remove the ffu_undercloud_repo_args/rhos_release bits
from defaults because they can be provided in the extra var from
outside instead.

Change-Id: I8284daebd4ddd5cb48ad2e2f510d688203dd8357