FFWD2: cleanup converge
We used to deal with docker-images.yaml file in FFWD1 but
now it only causes problems when we blindly reintroduce it.
This works for queens to train but fails on rocky to train
as the file was no longer used.

Change-Id: I8bc93fa88c53965f44e01393e7bd790924dc57f5
(cherry picked from commit b8668f4e085561f1f7f6da6f8eb889829ca95cd3)