Rename containers prepare parameters tag for ffu.
The ffu's containers prepare step was still using the old's upgrades
tag. This is an issue, as in infrared we skip steps by using skip_tags,
so as this task is wrongly tagged it ends up running when it shouldn't.

Change-Id: I5c57efc07d62bc1692626deb4a4808e21032f522
(cherry picked from commit fd02f2869d0b3fa116ef2bcb823125055927946b)