Add the -n flag in the run parameters conversion script
The FFU procedure will fail to adjust the NIC templates for environments with composable roles because it will point to the default network data file. In order to circumvent this, we want to provide a means to provide custom network data, which is done by adding the -n flag to the run parameters conversion script.

Change-Id: I50426d2cc296e237fa8a8cf6c750e760b0992dd2