Handle errors on system upgrade
Using 'wait' will not return error if any of the
child background process fails. This results in
unexpected behavior of triggering upgrade run
eventhough system upgrade failed. Observed this
issue during the integration of NFV's NIC partitioning
feature, in which system upgrade hangs on SR-IOV
cleanup step.

Used `wait $PID` to handle the states of all child
processes and wait for all to complete. Exit with error
if anyone of the childs exited with error.

Change-Id: I084f3b974e37c9f5a4d720747e408645deb50995