Add logic to execute validation groups in updates/upgrades.
This commit reshapes the existing updates logic to allow running validations
groups in a more generic way. Including a task file in which the following
is executed:
- Execute the group being passed validation_group if it's included in
  the validation_allowed_groups.
- Log the output into a log file.
- Retrieve the validations results and fail if there is any validation
  failed.
- Log all the tasks executed in each failed validation.

At the same time, this patch invokes the pre-upgrade validation group for
FFU, both before starting with the upgrade (which uses mistral to trigger
the validation in OSP13 [0]) and before starting with the overcloud upgrade
(which uses the new validations framework).

[0] - https://access.redhat.com/documentation/en-us/red_hat_openstack_platform/16.1/html/framework_for_upgrades_13_to_16.1/planning-and-preparation-for-an-in-place-openstack-platform-upgrade#validating-red-hat-openstack-platform-oldvernum-before-the-upgrade

Change-Id: I4c70a209283051b3e405bb24e03f555cf3802d94