Move FFU validations to right playbook.
In [0], the integration of validations was included. However, they
were added into the tas/upgrade/main.yaml playbook instead of
the tasks/fast-forward-upgrade/main.yml playbook, which is the
one being triggered during FFU automation.

Also, it adds back the rescue block in the validation_run_group.yaml
file. As otherwise, if a validation fails then the execution gets
stopped and never gets to run the check for failed validations. And
adds a check for the old release validations.

[0] - https://review.opendev.org/c/openstack/tripleo-upgrade/+/787411

Change-Id: I2fdd4c3c2c85482cf64f6565ad99cb4b71b00c13