[ffwd] Allow users to replace environment files
A new option allows user to replace the content of specific
environment files which are listed in the "overcloud deploy"
bash script which is parsed and used here.

This is useful when the content of a file is known to
have changed between the source and destination release.
The feature is implemented for FFWD only right now,
but it could be used for any kind of upgrade/update process.

Change-Id: I027770e0ac8c73896d2b57218da3a5063b9707fd