Rename unpredictable name interfaces in FFU.
Leapp does not work with unpredictable name interfaces (ethX naming).
Up to now, this limitation was being handled via workarounds [0].
However, it's time to apply the very same hack as it's being documented
in the FFU guideline [1].

This patch creates a new playbook based on the documented hack [1] and
executes it for the Undercloud and Overcloud nodes. In case any of the
nodes have any ethX interface name, a udev rule will be created and the
corresponding networks-script replace to match the new interface name.
A restar is required for the action to take place, but this will be
happening during the Leapp upgrade itself to avoid extra node restarts.

[0] - https://gitlab.cee.redhat.com/rhos-upgrades/workarounds/-/blob/master/upgrade/13-16.1_upgrade_workarounds.yaml#L3-14
[1] - https://access.redhat.com/documentation/en-us/red_hat_openstack_platform/16.1/html-single/framework_for_upgrades_13_to_16.1/index#using-predictable-nic-names-preparing-undercloud

Change-Id: I858f587536b34ac405276cafc9b328598ba73fc9