Add mechanism to skip certain validations.
When executing the tripleo-upgrade role with the option --run-validations
it might occur that a set of validations want to be skipped from the group
execution.

This patch adds a new parameter named skiplist_validations which contains
the list of validations names to be skipped. This parameter is also
integrated in the infrared plugin, by exposing the parameter
--skiplist-validations 'val1,val2,val3'

Change-Id: I47c1a2d0a8cc6618444b6301232079f69c5d8d6e