Take into account the validations return code.
Currently, if the execution of the validations fails due to some bug
or exception, this won't be cought. As the rescue block will be
triggered and as no validations failed (because non got to the point
of being executed) the check will be passed.

This patch registers the outcome from the command execution and verifies
that it was succeeded.

Change-Id: I7bac98032b84fadb6e42f8d790db3bba881dbc7b