[Train-Only] Parametrize ovs version to install in UC via leapp upgrade.
A new openvswitch version has been released, 2.15. As during the Undercloud
leapp upgrade we tell leapp which version of ovs to install in order to
save some extra restart, we now need to update the version to install.

However, the very same upstream branch is being used for two downstream
releases 16.1 and 16.2, and the openvswitch package version used on each
of them is different. For some reason, we need to parametrize it so that
we can specify which version to use for each CI job.

This patch adds a new parameter leapp_ovs_version which will determine
the openvswitch version to install while upgrading via leapp. By default,
it will use 2.13, which is the one used in 16.1.

Change-Id: I8647b36c0edc87e559dad748ced5c5acdeb91472
Related: rhbz#1966476