Add validations_result when running validations.
When merging [0], it was missed part of the solution submitted in [1]
removing the register: validations_result. As a consequence, if the
validations command fails then Ansible will show an error saying that
validations_result variable does not exist. This patch adds back
the parameter.
Also, it corrects the case when the validations_filtered parameter
does not contain any validation. In that case, the validation group
will be skipped.

[0] - I47c1a2d0a8cc6618444b6301232079f69c5d8d6e
[1] - I7bac98032b84fadb6e42f8d790db3bba881dbc7b

Change-Id: I18bea0f88926f47fe2945bd8c127d6a372efa52f
(cherry picked from commit bcdd0063749b2fbc74b131b389d760d1a630b420)