Set config_heat_extra_yaml before building converge script.
The converge script requires from the parameter config_heat_extra_yaml
to be set before we build the script, otherwise it won't add the
config_heat_extra.yaml file in the converge command.

This patch modifies the order, so common/build_extra_template.yaml is
executed before creating the script.

Change-Id: I31d6228720248908a66e05d5c145bad6a2c85312