Skip evacuations for non-compute nodes
It may happen that multiple hosts are passed into this
script, so we ensure that the grep OR's all the values.

We then also check if there are no hosts in the response,
and exit with failure if that is the case because this
script should always include at least one compute host.

We also revert the conditionals changed by:
https://review.opendev.org/773002

The conditions must be taken together to ensure that the
script is only created and run for computes.

Change-Id: I0d59e12d1523491618c83c0002edb8412e664be0
(cherry picked from commit 5cf94869ed02adcc56492dcac5aaa692304dfb59)