[Train-Only] Move the rename-nics after upgrade prepare.
During upgrade prepare in FFWD it looks like some of the stackrc parameters
get updated, therefore we need to update the tripleo-ansible-inventory
file using that environment file, as this inventory is being used to
execute the nics renaming on the overcloud nodes.

This patch adds a new task to regenerate the tripleo-ansible-inventory.yaml
file after running upgrade prepare and then it executes the nics renaming
script.

Change-Id: I4708ab72f742247a3a17318b2785862960de1724