Add different migration actions based on guest
There are currently workloads that do not support live-migration during
the upgrades process. Example workloads include guests with SR-IOV
ports, pci-passthrough, ovs+dpdk vhost ports, or cpu pinning. Nova would
also like to test the results of upgrades when a workload is not
migrated off a host and is left in place.

This commit checks a workloads metadata and takes the appropriate action
based on the data. I.e. it will live, cold, or not migrate based on the
metadata of the instance. If no meta-data is present on the instance
describing the workload action, it will default to live migration unless
compute_cold_evacuate is set, in which case it will default to cold
migration.

Change-Id: I781a894089320067d0ad3e174aed18fabe7c6ce2