Move the upgrade prepare script generation in the build script playbook.
Currently, the overcloud_upgrade_prepare.sh script was being built in the
overcloud_upgrade_prepare.yaml playbook to make sure the upgrade_prepare_extra_params
parameter was up to date. However, this parameter is not used anymore in
the ffu workflow and the upgrades workflow is not being used by our CI so
it should be safe to move it back to the create-overcloud-ffu-script.yaml.

Change-Id: I32494fb21d40cd3231e619c748fdfce7ff891f50