Make the gerrit_url of patches configurable
Currently patches must come from review.openstack.org but it may
be useful to also source them from an alternative gerrit.

This patch adds a new parameter called 'gerrit_url' which is the
URL to the gerrit where the patch should be sourced from. The
default value is 'https://review.opendev.org' to ensure that it
maintains the current behaviour for backwards compatibility.

Change-Id: I84f2a7b9aae674d1971b4f64628ca59bf40419ad