Bump the default RHCS 3 tag to the newest (3-51)
The new release include several fixes, and using a lower version
than the one available in the source system may lead to issues.
For example, a bug in 3-40 prevents the restarting of ceph-nfs
containers (see https://bugzilla.redhat.com/show_bug.cgi?id=1801090).

In general it should be possible to override the ceph parameters
through the infrared plugin as well.

Change-Id: I8c834c59cb881d739399ded56864ce4f74e1c8a8