Ensure migrate all the VMs before upgrade the compute
Currently, there is one case that escape from the nest if/else
statements. When there are instances to migrate and there is no
DO_NO_MIGRATE_IDs, there was returing 0.

This patch fixed that and makes the script return 1, because
there are VMs to migrate.

This patch also removed the second "host_quiesced" called, as
it was previously executed on the while statement.

Co-authored-by: Daniel Mats Niklas Bengtsson <dbengt@redhat.com>
Change-Id: Idfb9899cc9a599bff2e335c74edd7ee4f74e8c06
(cherry picked from commit 305d9fb0d830fdd98b7dc4c4bd92b833ed5916d4)