Make a workload_launch script more robust
Add more error detection in workload_launch script to fail script early
when the error occurs.

Also, make ssh test to run by default to be sure created vm is
accessible via floating ip, before saving its floating ip to a file. The
file is used in ping test and we need to make sure it is not present on
undercloud when vm fails to boot or is unreachable.

Change-Id: I1057de8dde62fef7e2e00880704d7f3bb3bf6b19