Create entry for empty groups in inventory_rolemap
When a group is a child of another group then the child groups role mappings
are added to the parent group role mappings. This which fails with a
KeyError when the child is an empty group as no mapping is created.

Conflicts:
	filter_plugins/tripleo_upgrade.py

Closes-bug: #1994081
Related: rhbz#2089512
Change-Id: I05f4989b0f14d56d3c33f91b3cccb85b729987cb
(cherry picked from commit 2eea005773a5d9b2a32c88e1427d839788d00dd6)