[workload_launch] Update µversions for Train & Victoria
The microversions are key to ensuring that the workload
testing works whether the Undercloud has been upgraded
or not. These are set to ensure that we do all workload
testing at the source version.

Microversions are updated for:

 - Nova [a]
 - Cinder [b]
 - Manila [c]

[a] https://docs.openstack.org/nova/latest/reference/api-microversion-history.html
[b] https://docs.openstack.org/cinder/latest/contributor/api_microversion_history.html
[c] https://docs.openstack.org/manila/latest/contributor/api_microversion_history.html

Backport note:
The versions applied are different from the original cherry-pick due to the different source versions for this branch.

Change-Id: I8fe9727570c193f96de99b812f223c7ca7dc4761
(cherry picked from commit ca284f8785186317a386a2e2ab6a0c102026e924)