Remove centos8 wallaby jobs from zuul layouts
This removes the remaining c8 check and gate jobs from running for
stable/wallaby. We are removing wallaby/c8 integration and component
lines in [1]. For c8/wallaby we are keeping only the ffu job so
this patch removes the c8 minor update job and content provider

See [2] for related patches.

[1] https://review.rdoproject.org/r/q/topic:oooci_teardown_8_wallaby
[2] https://review.opendev.org/q/topic:oooci_teardown_8_wallaby

Change-Id: I361df71622fe2edee94d6f621c02030770aec212