workload launch refactor to support continuous control plane test.
This refactor paves the way to support continuous control plane test.

Added:
- trigger error when openstack command fails during creation;
- add a trap exit to make sure cleanup is always trigger in sanity;
- add a sanityfast mode to fail fast during continuous testing;
- add more test on created resources like cinder vol and fip;
- remove leftover ~/instance_.. file during sanity;

Sanity mode, we add a trap at exit to support cleanup. This ensures
that cleanup is triggered if any openstack command fail so that we
don't need explicit test.

During the teardow, we don't trigger new error for openstack command
as the state of the overcloud is at that point unknown and the best
course of action is just to return an error at that point.

Change-Id: I1e15ccb5a13e253da545efb6b1a23e4b20ef3faa
(cherry picked from commit 1dda7e142eade0035eef141b2c2815e2df3b2d02)