Add control plane testing during update.
continous-test.sh is a light shell daemon to start "workload_launch.sh
sanityfast" in a loop.

That daemon collects the log of each individual run in a "ct-<PID>"
directory and the result in control-plane-test-<PID>.log.  The pid
can be found in control-plane-test.pid.

In tripleo-upgrade we add the usual check integration at the relevant
part of the update. The new boolean option
"--update-controlplane-check" is added to the infrared interface.

By default this test is not activated. control_plane_check variable
controls that behavior outside of infrared.

We have two definitions of failure:
- how many successive failures we have;
- how many total failures we have.

This should avoid false positive while still catching long enough cut.

We also activate debug output for the workload launch script to ease
error checking.

Change-Id: Ieff860667cf67a1c1f67221b05575e952a0636fa