Add support for all arguments on one line for deploy script.
For update prepare we need to find back the parameters from the
deployment.

The problem is that the current awk one liner expect a specific format
for the command line: only one parameter followed by its value on one
line. If the all the parameters are passed on one line, then the
script breaks.

Add support for both formats (one paramater per line and all parameter
on one line).

We squeeze all spaces, then convert them to newlines and parse the
resulting file using awk.

Change-Id: Icd8b247447c805b9a102d94fbd6ee7019552c87c