[update] Add ha-image-update external run.
The templates now integrate a test to verify very early if we need
container image's name under pacemaker control need to be modified[1].

When this happen we need to run

    openstack overcloud external-update run --tags ha_image_update

before overcloud update run on the Controller to ensure the control
plane doesn't break during update.

This patch offer the possibility to trigger this command when the
operator know that there will be such change and thus will prevent the
check from[1] to fail the update.

We also collect the log before that stage.

[1] https://review.opendev.org/c/openstack/tripleo-heat-templates/+/861100/
    look for the "Check for update of * container image name" tasks

Change-Id: I685e8ef117a394629b9be4bb6dfa8f6301bf6e89