[ffwd] Modify network environement instead of drop
We used to remove network-environement.yaml file as we get the
network environment data from UC export. Sadly this file is often
also used to pass other parameters than just network environment.

To solve this and resolve resource conflict we simply reuse the file
but drop the resource_registry from it.

Change-Id: I9369c453fb5ec7acf9ace70d3d308c652c9ee547