Only manage network-environment content when the file exits
Currently, the plugin try to modify the network-environment file
as long as the file is defined. This cause an issue with the
default environment file
"/usr/share/openstack-tripleo-heat-templates/environments/network-environment.yaml"
as this file does not exit and it uses the .j2.yaml to generated
"/usr/share/openstack-tripleo-heat-templates/environments/network-environment.j2.yaml"

This patch only tries to modify the network-environment file
when the file exits.

Change-Id: I38d8c62996c0065669a96370764ce36d0cd0600c