workload : generate traffic to the vm.
Some issues on the dataplane are more likely to happen when there have
been some traffic to the vm. Furthermore it match more closely what
customer would have (running vms with existing traffic)

This patch add a run of iperf during 1 minute to simulate that traffic
just after the vm has been created.

This is opt-in as the mode needs to be changed to workload_traffic. To
access it add `--install-upgrade-workload-traffic` to the infrared
command or set `workload_launch_traffic` to true in other
environments.

We also take into account vm with misconfigured name resolution, by
temporarily taking the resolv.conf from the undercloud (which should
be always working) and use that for getting iperf into the vm.

Change-Id: Ide4d840ba8e42ef57b7bf748a65307d4ed5cc9a6