Remove the systemd status command from the log collection.
The `status` command can return codes different than 0 that
just mean that a service is in transition, like for instance, when activating a service from a timer. For instance:

    Active: activating (start)

will return code 3, as do:

   Active: failed

So, that means that we should ignore this return in all case unless we
do a specific test for service in fail state.

As we get all the unit files with the previous log collection, which
use `list-units`, this `status` run is useless.

If we still want the check, we have to add another review which will
explicitly check for failed services:

    systemctl status --failed tripleo*

Change-Id: I80afbe7a79fc6ce923061f0397fa71f431bad68e