Skip check-reboot from post update validation.
check-reboot will report a failure if reboot is needed, which is often
the case after update.

Just skip this informative check to avoid ci failure.

We add a new variable because skiplist_validations will be overwriten
by infrared. So to be able to define a hardcoded check to skip at the
stage level, we need a brand new variable.

Change-Id: I2248263386b9d581c2e75707d0c9fdf1fb888355