Change registered variable name in 'register ntp-server opts' task
To prevent naming conflicts ntp_server variable registered in 'register
ntp-server opts' task has to be changed to something unique. ntp_server
is defined in tripleo-ci repos and cannot be re-used.

This solves error we hit in downstream CI:

`The conditional check 'ntp_server is succeeded' failed. The error was:
The 'failed' test expects a dictionary`

Change-Id: I9be3eed5d462b9a195416162166b82324edbabfd