Remove --force on the patch checking
Currently, trying to apply the patch [1] on the CI is failing
becuase the patch was already applied, and on the second
execution the patch fail and make the gate failed.

This patch try to fix that, and do the check in the same
way as the patch is going to be applied.

Change-Id: I5ade3f6a22ea60db1e914c7bc1a9578557d5342c